+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Push tags to *aliyun* repository.
+- Move `rbac` controller code into `rbac` package.
 
 ## [0.3.3] - 2020-05-13