+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+##[Unreleased]
+
+
+## [0.4.0]
 
 ### Changed
 
@@ -49,7 +52,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Read-only role for customer access into Control Plane.
 
-[Unreleased]: https://github.com/giantswarm/rbac-operator/compare/v0.3.3...HEAD
+[Unreleased]: https://github.com/giantswarm/rbac-operator/compare/v0.4.0...HEAD
+
+[0.4.0]: https://github.com/giantswarm/rbac-operator/releases/tag/v0.4.0
 
 [0.3.3]: https://github.com/giantswarm/rbac-operator/releases/tag/v0.3.3